Refactor Phase 2
