+++ app/api/post/model.py
@@ -1,4 +1,4 @@
-from core import db
+from app import db
 
 
 class Post(db.Model):
