+++ app/api/user/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+user = Blueprint('user', __name__)
+
+from . import controller
